Compatibility with newer Hugo versions (#2298)
* remove pointless bt_wiki

* replace .URL with .Permalink/.RelPermalink

.URL has been deprecated for several years now, and Hugo suggests
using .RelPermalink instead.

In places where we use the url only for comparisons, it doesn't really
matter if it's .Permalink or .RelPermalink - as long as both sides
of the comparison are the same

* In templates, use the "hugo" variable

...rather than the deprecated .Hugo

* Switch the markdown renderer to goldmark

New hugo versions (eg. 0.101) drop support for blackfriday, and they
suggest goldmark instead.

Seemingly the only change that was needed was in this shortcode:
without removing whitespace, goldmark inserted an additional <p>,
which breaks the layout.

* readme: note on local development with docker

* Correct markdown link syntax

The syntax is `[label](link)`, not `[label] (link)`. Whitespace is
not allowed, and I don't know why blackfriday allowed it.