RD-5110 The new `heal` workflow (#2319) (#2416)
* RD-5110 The new `heal` workflow

It's a bit more involved than it used to be - it's not about just
reinstalling anymore!

Also, I'm removing those images. The graph isn't needed there, and it's
bad enough that it makes the example harder to understand.

Instead, I simplify the blueprint a bit. No need for `database_host`.

* description: s/Heals/Heal/

* Explain `node_instance_id` additionally

Co-authored-by: Lukasz Maksymczuk <tehasdf@users.noreply.github.com>