RD-5110 The new heal workflow (#2417)
* Snyk latest (#2414)

* fix: upgrade mustache from 2.2.1 to 2.3.2 (#2003)

Snyk has created this PR to upgrade mustache from 2.2.1 to 2.3.2.

See this package in npm:
https://www.npmjs.com/package/mustache

See this project in Snyk:
https://app.snyk.io/org/alex-5fb/project/95c123bb-d16b-43b2-a748-3e6bd1c7fccc?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: alexmolev <amolev@gmail.com>

* fix: upgrade express from 4.13.4 to 4.17.1 (#2002)

Snyk has created this PR to upgrade express from 4.13.4 to 4.17.1.

See this package in npm:
https://www.npmjs.com/package/express

See this project in Snyk:
https://app.snyk.io/org/alex-5fb/project/95c123bb-d16b-43b2-a748-3e6bd1c7fccc?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: alexmolev <amolev@gmail.com>

* fix: upgrade socket.io from 1.3.7 to 1.7.4 (#2001)

Snyk has created this PR to upgrade socket.io from 1.3.7 to 1.7.4.

See this package in npm:
https://www.npmjs.com/package/socket.io

See this project in Snyk:
https://app.snyk.io/org/alex-5fb/project/95c123bb-d16b-43b2-a748-3e6bd1c7fccc?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: alexmolev <amolev@gmail.com>

Co-authored-by: Snyk bot <github+bot@snyk.io>

* RD-5110 The new `heal` workflow (#2319)

* RD-5110 The new `heal` workflow

It's a bit more involved than it used to be - it's not about just
reinstalling anymore!

Also, I'm removing those images. The graph isn't needed there, and it's
bad enough that it makes the example harder to understand.

Instead, I simplify the blueprint a bit. No need for `database_host`.

* description: s/Heals/Heal/

* Explain `node_instance_id` additionally

Co-authored-by: Snyk bot <github+bot@snyk.io>
Co-authored-by: Lukasz Maksymczuk <tehasdf@users.noreply.github.com>